requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
black==24.10.0
    # via -r requirements.in
click==8.1.7
    # via black
ffmpeg-python==0.2.0
    # via -r requirements.in
flake8==7.1.1
    # via -r requirements.in
future==1.0.0
    # via ffmpeg-python
isort==5.13.2
    # via -r requirements.in
mccabe==0.7.0
    # via flake8
mypy-extensions==1.0.0
    # via black
packaging==24.1
    # via black
pathspec==0.12.1
    # via black
platformdirs==4.3.6
    # via black
pycodestyle==2.12.1
    # via flake8
pyflakes==3.2.0
    # via flake8
